git commit -m “[feat] offer Dto, Entity의 Id관련 부분 Intger -> Long으로 변경

+++ src/main/java/com/example/lol_likelion/duo/dto/OfferDto.java
@@ -15,11 +15,11 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class OfferDto {
 
-    private Integer id;
+    private Long id;
     @Setter
     private String status;
     @Setter
-    private Integer applyUserId;
+    private Long applyUserId;
     @Setter
     private LocalDateTime createdAt;
     @Setter
